+++ R/fps.R
@@ -332,18 +332,32 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
                 tth_align, align, manual_align, tth_fps, shift,
                 remove_trace, ...) {
 
+#---------------------------------------------------
+#Conditions
+#---------------------------------------------------
+
+#Make sure there aren't any negative counts
+  if (min(smpl[[2]]) < 0) {
+
+    stop("Please make sure that there are no negative count intensities in the sample data")
+
+  }
+
+#Set harmonise = TRUE as the default if missing
   if (missing(harmonise)) {
 
     harmonise <- TRUE
 
   }
 
+#Make sure harmonise is logical
   if (!is.logical(harmonise)) {
 
     stop("The harmonise argument must be logical.")
 
   }
 
+#Make sure that the user knows to use harmonise if needed
   if (harmonise == FALSE & !identical(lib$tth, smpl[[1]])) {
 
     stop("The 2theta scale of the library and sample do not match. Try
@@ -351,18 +365,21 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
 
   }
 
+#Set the std_conc to NA if missing
   if (missing(std_conc)) {
 
     std_conc <- NA
 
   }
 
+#Make sure that the std_conc is numeric if needed
   if (!(is.numeric(std_conc) | is.na(std_conc))) {
 
     stop("\n-The std_conc argument must either be NA or a numeric value greater than 0 and less than 100.")
 
   }
 
+#Ensure the std argument is defined if needed and that it is between 0 and 100
   if (is.numeric(std_conc)) {
 
     if (missing(std)) {
@@ -392,12 +409,14 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
     align = 0.1
   }
 
+#Set the default alignment type to automated
   if(missing(manual_align)) {
 
     manual_align <- FALSE
 
   }
 
+#Make sure manual_align is logical
   if(!is.logical(manual_align)) {
 
     stop("The manual_align argument must be logical")
@@ -423,7 +442,7 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
     remove_trace = 0
   }
 
-  #If refs are not defined or "all" then use all of them
+  #If refs are not defined then use all of them
   if(missing(refs)) {
 
     cat("\n-Using all reference patterns in the library")
@@ -456,7 +475,7 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
     warning("Be cautious of large 2theta shifts. These can cause issues in sample alignment.")
   }
 
-  #Create a warning message if the shift is greater than 0.5, since this can confuse the optimisation
+  #Ensure that remove_trace is not less than zero
   if (remove_trace < 0) {
     stop("The remove_trace argument must be greater than 0.")
   }
@@ -466,14 +485,17 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
     stop("The solver argument must be one of 'BFGS', 'Nelder Mead', 'CG', or 'NNLS'")
   }
 
+  #Use "BFGS" instead of "L-BFGS-B" if defined
   if (solver == "L-BFGS-B") {
 
     cat("\n-The 'L-BFGS-B' option for solver has deprecated.
         Using 'BFGS' instead.")
 
+    solver <- "BFGS"
+
   }
 
-  #If the solver is "NNLS" and shift > 0, then stop
+  #If the solver is "NNLS" and shift > 0, then don't shift
   if (solver == "NNLS" & shift > 0) {
 
     warning("When shift > 0, the solver argument must be one of 'BFGS',
@@ -494,12 +516,12 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
 
   }
 
+#-------------------------------------------------------------
+#END OF CONDITIONS, NOW SUBSET LIBRARY
+#-------------------------------------------------------------
 
-
-  #subset lib according to the phases vector
-
-  lib$xrd <- lib$xrd[which(lib$phases$phase_id %in% refs | lib$phases$phase_name %in% refs)]
-  lib$phases <- lib$phases[which(lib$phases$phase_id %in% refs | lib$phases$phase_name %in% refs), ]
+  #subset lib according to the refs vector
+  lib <- subset(lib, refs = refs, mode = "keep")
 
   #Make sure that the phase identified as the internal standard is contained within the reference library
   if (!std == "none" & !std %in% lib$phases$phase_id) {
@@ -516,6 +538,9 @@ fps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, std_conc,
 
   }
 
+#--------------------------------------------------------
+#Alignment
+#--------------------------------------------------------
 
 if (!align == 0) {
 
@@ -579,49 +604,25 @@ lib$xrd <- lib$xrd[which(lib$tth >= tth_fps[1] & lib$tth <= tth_fps[2]), , drop 
 #Replace the tth in the library with the shortened one
 lib$tth <- smpl[, 1]
 
-if (solver %in% c("Nelder-Mead", "BFGS", "CG")) {
 
-#--------------------------------------------
-#Initial Optimisation
-#--------------------------------------------
+#-----------------------------------------------------------
+#Initial optimisation
+#-----------------------------------------------------------
+
+#The optimisation can fail if negative have creeped in during interpolation
+if(length(which(smpl[[2]] < 0) > 0)) {
+
+  delete_negs <- which(smpl[[2]] < 0)
+  smpl <- smpl[-delete_negs,]
+  lib$tth <- lib$tth[-delete_negs]
+  lib$xrd <- lib$xrd[-delete_negs, ]
+
+}
+
+if (solver %in% c("Nelder-Mead", "BFGS", "CG")) {
 
 x <- rep(0, ncol(lib$xrd))
 names(x) <- names(lib$xrd)
-
-#Optimise weighting and shifts if these conditions are met
-#if (shift_mode == "optimise" & shift > 0 & length(x) > 1) {
-
-#Create x_s parameters
-#x_s <- x
-#names(x_s) <- paste0(names(x), "_s")
-
-#cat("\n-Optimising weighting and shifting coefficients...")
-#o <- stats::optim(par = c(x,x_s), .fullpat_shift_optim,
-#                  method = solver, lib = lib,
-#                  smpl = smpl, obj = obj)
-
-#x <- o$par
-
-#Extract the shifted data
-#cat("\n-Harmonising library and sample to same 2theta axis")
-#shifted <- .fullpat_shift(smpl = smpl, lib = lib,
-#                          par_shift = x[((length(x)/2)+1):length(x)],
-#                          limit = shift)
-
-#lib <- shifted$lib
-#smpl <- shifted$smpl
-
-#Now extract just the weighting coefficients
-#x <- x[1:ncol(lib$xrd)]
-
-#Now make sure negative coefficients are (almost) zero again
-#if (length(which(x < 0)) > 0) {
-
-#  x[which(x < 0)] <- 1*10^-16
-
-#}
-
-#} else {
 
 #Initial optimisation
 cat("\n-Optimising...")
@@ -632,11 +633,10 @@ o <- stats::optim(par = x, .fullpat,
 
 x <- o$par
 
-#}
 
-#-----------------------------------------------
+#------------------------------------------------------------
 # Remove negative parameters or parameters equal to zero
-#-----------------------------------------------
+#------------------------------------------------------------
 
 if (min(x) < 0) {
 
@@ -667,17 +667,6 @@ lib <- remove_neg_out[[2]]
 #is greater than zero and the correct solver arguments are used
 
 if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
-
-  #fpf_aligned <- .shift(smpl = smpl,
-  #                      lib = lib,
-  #                      max_shift = shift,
-  #                      x = x,
-  #                      res = shift_res,
-  #                      obj = obj)
-
-  #smpl <- fpf_aligned[["smpl"]]
-  #lib$xrd <- data.frame(fpf_aligned[["lib"]])
-  #lib$tth <- smpl[,1]
 
   #This will replace the grid search shifting
   cat("\n-Optimising shifting coefficients...")
@@ -710,9 +699,17 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
 #Re-optimise after shifting
 #----------------------------------------------
 
-  #if (solver %in% c("Nelder-Mead", "BFGS", "CG")) {
-
     cat("\n-Reoptimising after shifting data")
+
+  #The optimisation can fail if negative have creeped in during interpolation
+  if(length(which(smpl[[2]] < 0) > 0)) {
+
+    delete_negs <- which(smpl[[2]] < 0)
+    smpl <- smpl[-delete_negs,]
+    lib$tth <- lib$tth[-delete_negs]
+    lib$xrd <- lib$xrd[-delete_negs, ]
+
+  }
 
     o <- stats::optim(par = x, .fullpat,
                       method = solver, pure_patterns = lib$xrd,
@@ -720,19 +717,6 @@ if(shift > 0 & length(x) > 1 & solver %in% c("Nelder-Mead", "BFGS", "CG")) {
 
     x <- o$par
 
-  #}
-
-  #else {
-
-  #  cat("\n-Applying non-negative least squares")
-
-  #  nnls_out <- .xrd_nnls(xrd.lib = lib, xrd.sample = smpl[, 2])
-
-  #  lib$xrd <- nnls_out$xrd.lib
-
-  #  x <- nnls_out$x
-
-  #}
 
 #------------------------------------------------------
 #Remove negative/zero parameters after shifting